Improve malformed-input handling, restarting, and crash-recovery
FTW v0.48 was recently shipped, fixing a crash that could happen when the
webserver was sent an `HTTP/0.9` request (what is this, 1991!?), and fixing
a scenario where the webserver would fail to return when stopped.

Also properly overrides `LogStash::Inputs::Base#stop`, so we can ensure our
server gets stopped when the plugin is stopped or reloaded.